Identifiers for variables in the Hector carbon cycle component

These identifiers correspond to variables that can be read and/or set in the carbon cycle

Usage

NBP()

CONCENTRATIONS_CO2()

NPP()

RH()

RH_DETRITUS()

RH_SOIL()

ATMOSPHERIC_CO2()

FFI_EMISSIONS()

DACCS_UPTAKE()

LUC_EMISSIONS()

LUC_UPTAKE()

CO2_CONSTRAIN()

NBP_CONSTRAIN()

VEG_C(biome = "")

DETRITUS_C(biome = "")

SOIL_C(biome = "")

PERMAFROST_C(biome = "")

THAWEDP_C(biome = "")

FRAC_FROZEN(biome = "")

FRAC_STATIC(biome = "")

PERMAFROST_MU(biome = "")

PERMAFROST_SIGMA(biome = "")

FRAC_DECOMP_CH4(biome = "")

EARTH_C()

NPP_FLUX0(biome = "")

Arguments

biome

Name of biome (leave empty for global)

Functions

CO2_CONSTRAIN(): Constrain atmospheric CO2 concentration ("(ppmv CO2)")

NBP_CONSTRAIN(): Constrain net biome production (land-atmosphere flux) ("(PgC/yr)")

VEG_C(): Vegetation C pool (`"Pg C"`)

DETRITUS_C(): Vegetation detritus C pool (`"Pg C"`)

SOIL_C(): Soil C pool (`"Pg C"`)

PERMAFROST_C(): Permafrost C pool (`"Pg C"`)

THAWEDP_C(): Thawed permafrost C pool (`"Pg C"`)

FRAC_FROZEN(): Fraction of permafrost still frozen ("(unitless)")

FRAC_STATIC(): Fraction of thawed permafrost that is static ("(unitless)")

PERMAFROST_MU(): Permafrost thaw mu parameter ("(unitless)")

PERMAFROST_SIGMA(): Permafrost thaw sigma parameter ("(unitless)")

FRAC_DECOMP_CH4(): Methane fraction of permafrost decomposition ("(unitless)")

NPP_FLUX0(): Initial net primary productivity (NPP)flux (`"Pg C year^-1"`)

Note

See Woodard et al. (2021)

See Woodard et al. (2021)

Output variables

These variables can be read using the GETDATA message type:

CONCENTRATIONS_CO2

Atmospheric CO2 concentration

ATMOSPHERIC_CO2

Atmospheric CO2 in units of C

NBP

Net biome production. Annual global C flux from atmosphere into the land. A positive value means a net flux from atmosphere into land (i.e. land is a net carbon sink), while a negative value means a net flux from land into the atmosphere (i.e. land is a net carbon source). Note that despite the name–which follows disciplinary convention, see Chapin et al. 2006–this is currently a globally-averaged variable.

FFI_EMISSIONS

Fossil fuel and industrial emissions

LUC_EMISSIONS

Land use change emissions

NPP

Net primary production

RH

Heterotrophic respiration

EARTH_C

Earth pool

Input variables

FFI_EMISSIONS

Fossil fuel and industrial emissions ("Pg C/yr")

LUC_EMISSIONS

Land use change emissions ("Pg C/yr")

CO2_CONSTRAIN

Prescribed atmospheric CO2 concentration ("ppmv CO2")

NBP_CONSTRAIN

Prescribed net biome production (land-atmosphere C flux) ("Pg C/yr")

DACCS_UPTAKE

Direct air carbon capture and storage

LUC_UPTAKE

Land use change uptake

Note

Because these identifiers are provided as #define macros in the hector code, these identifiers are provided in the R interface as functions. Therefore, these objects must be called to use them; e.g., GETDATA() instead of the more natural looking GETDATA.

See also