Retrieve the current list of biomes for a Hector instance

Retrieve the current list of biomes for a Hector instance

Usage

get_biome_list(core)

Arguments

core

Handle to the Hector instance from which to retrieve the biome list.