Identifiers for variables in the Hector carbon cycle component

These identifiers correspond to variables that can be read and/or set in the carbon cycle

LAND_CFLUX()

ATMOSPHERIC_CO2()

ATMOSPHERIC_C()

FFI_EMISSIONS()

LUC_EMISSIONS()

Output variables

These variables can be read using the GETDATA message type:

ATMOSPHERIC_CO2

Atmospheric CO2 concentration

ATMOSPHERIC_C

Atmospheric carbon pool

LAND_CFLUX

Land carbon flux

FFI_EMISSIONS

Fossil fuel and industrial emissions

LUC_EMISSIONS

Land use change emissions

Input variables

FFI_EMISSIONS

Fossil fuel and industrial emissions ("Pg C/yr")

LUC_EMISSIONS

Land use change emissions ("Pg C/yr")

Note

Because these identifiers are provided as #define macros in the hector code, these identifiers are provided in the R interface as function. Therefore, these objects must be called to use them; e.g., GETDATA() instead of the more natural looking GETDATA.

See also