Identifiers for miscellaneous emissions not elsewhere described

All of these variables can be set using the SETDATA message type.

EMISSIONS_BC()

EMISSIONS_N2O()

EMISSIONS_NOX()

EMISSIONS_CO()

EMISSIONS_NMVOC()

EMISSIONS_OC()

Functions

EMISSIONS_BC: Black carbon emissions ("Tg")

EMISSIONS_N2O: N2O emissions ("Tg N")

EMISSIONS_NOX: Emissions for NOx compounds ("Tg N")

EMISSIONS_CO: Emissions for carbon monoxide ("Tg CO")

EMISSIONS_NMVOC: Emissions for non-methane volatile organic compounds (NMVOC) ("Tg NMVOC")

EMISSIONS_OC: Emissions for organic carbon ("Tg")

Note

Because these identifiers are provided as #define macros in the hector code, these identifiers are provided in the R interface as function. Therefore, these objects must be called to use them; e.g., GETDATA() instead of the more natural looking GETDATA.

See also