Identifiers for capabilities in the Hector forcing component

These identifiers specify forcing values that can be provided by hector via the forcing component. All of the values corresponding to these identifiers are read-only (i.e., they can only appear in GETDATA messages.)

RF_TOTAL()

RF_T_ALBEDO()

RF_CO2()

RF_N2O()

RF_H2O_STRAT()

RF_O3_TROP()

RF_BC()

RF_OC()

RF_SO2D()

RF_SO2I()

RF_SO2()

RF_VOL()

RF_CH4()

Functions

RF_TOTAL: Total radiative forcing

RF_T_ALBEDO: Radiative forcing due to albedo

RF_CO2: Radiative forcing due to CO2

RF_N2O: Radiative forcing due to N2O

RF_H2O_STRAT: Radiative forcing due to stratospheric water vapor

RF_O3_TROP: Radiative forcing due to tropospheric ozone

RF_BC: Radiative forcing due to black carbon

RF_OC: Radiative forcing due to organic carbon

RF_SO2D: Direct contribution of SO2 to radiative forcing

RF_SO2I: Indirect contribution of SO2 to radiative forcing

RF_SO2: Total contribution (direct and indirect) of SO2 to radiative forcing

RF_VOL: Radiative forcing due to volcanic activity

RF_CH4: Radiative forcing due to methane

Note

Because these identifiers are provided as #define macros in the hector code, these identifiers are provided in the R interface as function. Therefore, these objects must be called to use them; e.g., GETDATA() instead of the more natural looking GETDATA.

See also