Identifiers for quantities in the ocean component

These identifiers correspond to variables that can be read from the ocean component using the GETDATA message type.

Usage

OCEAN_UPTAKE()

OCEAN_C()

OCEAN_C_HL()

OCEAN_C_LL()

OCEAN_C_ML()

OCEAN_C_IO()

OCEAN_C_DO()

OCEAN_PREIND_C_SURF()

OCEAN_PREIND_C_ID()

TT()

TU()

TWI()

TID()

PH_HL()

PH_LL()

PH()

HL_OCEAN_UPTAKE()

LL_OCEAN_UPTAKE()

PCO2_HL()

PCO2_LL()

PCO2()

DIC_HL()

DIC()

DIC_LL()

SST_HL()

SST_LL()

CO3_LL()

CO3_HL()

CO3()

Functions

OCEAN_UPTAKE(): Atmosphere-ocean carbon flux

OCEAN_C(): Ocean total carbon pool

OCEAN_C_HL(): Ocean surface high-latitude carbon pool

OCEAN_C_LL(): Ocean surface low-latitude carbon pool

OCEAN_C_ML(): Ocean surface or mixed layer carbon pool

OCEAN_C_IO(): Intermediate ocean carbon pool

OCEAN_C_DO(): Deep ocean carbon pool

OCEAN_PREIND_C_SURF(): Preindustrial surface ocean carbon pool

OCEAN_PREIND_C_ID(): Preindustrial intermediate+deep ocean carbon pool

TT(): Thermohaline overturning

TU(): High-latitude overturning

TWI(): Warm-intermediate exchange

TID(): Intermediate-deep exchange

PH_HL(): High-latitude pH

PH_LL(): Low-latitude pH

PH(): Ocean surface pH

HL_OCEAN_UPTAKE(): Atmosphere-ocean carbon flux, high-latitude

LL_OCEAN_UPTAKE(): Atmosphere-ocean carbon flux, low-latitude

PCO2_HL(): Partial pressure of CO2, high-latitude

PCO2_LL(): Partial pressure of CO2, low-latitude

PCO2(): Ocean surface partial pressure of CO2

DIC_HL(): Dissolved inorganic carbon, high-latitude

DIC(): Ocean surface dissolved inorganic carbon

DIC_LL(): Dissolved inorganic carbon, low-latitude

SST_HL(): Absolute ocean surface temperature, high-latitude (deg C)

SST_LL(): Absolute ocean surface temperature, low-latitude (deg C)

CO3_LL(): Carbonate concentration, low-latitude

CO3_HL(): Carbonate concentration, high-latitude

CO3(): Ocean surface carbonate concentration

Note

Because these identifiers are provided as #define macros in the hector code, these identifiers are provided in the R interface as functions. Therefore, these objects must be called to use them; e.g., GETDATA() instead of the more natural looking GETDATA.

See also