Identifiers for quantities in the ocean component

These identifiers correspond to variables that can be read from the ocean component using the GETDATA message type.

OCEAN_CFLUX()

OCEAN_C()

OCEAN_C_HL()

OCEAN_C_LL()

OCEAN_C_IO()

OCEAN_C_DO()

TT()

TU()

TWI()

TID()

PH_HL()

PH_LL()

ATM_OCEAN_FLUX_HL()

ATM_OCEAN_FLUX_LL()

PCO2_HL()

PCO2_LL()

DIC_HL()

DIC_LL()

TEMP_HL()

TEMP_LL()

CO3_LL()

CO3_HL()

Functions

OCEAN_CFLUX: Atmosphere-ocean carbon flux

OCEAN_C: Ocean total carbon pool

OCEAN_C_HL: Ocean surface high-latitude carbon pool

OCEAN_C_LL: Ocean surface low-latitude carbon pool

OCEAN_C_IO: Intermediate ocean carbon pool

OCEAN_C_DO: Deep ocean carbon pool

TT: Thermohaline overturning

TU: High-latitude overturning

TWI: Warm-intermediate exchange

TID: Intermediate-deep exchange

PH_HL: High-latitude Ph

PH_LL: Low-latitude Ph

ATM_OCEAN_FLUX_HL: Atmosphere-ocean carbon flux, high-latitude

ATM_OCEAN_FLUX_LL: Atmosphere-ocean carbon flux, low-latitude

PCO2_HL: Partial pressure of CO2, high-latitude

PCO2_LL: Partial pressure of CO2, low-latitude

DIC_HL: Dissolved inorganic carbon, high-latitude

DIC_LL: Dissolved inorganic carbon, low-latitude

TEMP_HL: Ocean temperature, high-latitude

TEMP_LL: Ocean temperature, low-latitude

CO3_LL: Carbonate concentration, low-latitude

CO3_HL: Carbonate concentration, high-latitude

To Do

Some of these could be made writeable as well as readable. Notably, TT, TU, TWI, and TID are only set from the input, not calculated.

Note

Because these identifiers are provided as #define macros in the hector code, these identifiers are provided in the R interface as function. Therefore, these objects must be called to use them; e.g., GETDATA() instead of the more natural looking GETDATA.

See also