Shutdown a hector instance

Shutting down an instance will free the instance itself and all of the objects it created. Any attempted operation on the instance after that will raise an error.

shutdown(core)

Arguments

core

Handle to a Hector instance

Value

The Hector instance handle

Caution

This function should be called as mycore <- shutdown(mycore) so that the change from active to inactive will be recorded in the caller.

See also